Moved graphics to specific folder
