+++ src/models/graphics/Curve.ts
@@ -1,8 +1,8 @@
 import * as SVG from "svg.js";
-import Utilities from "../utilities/general";
+import Utilities from "../../utilities/general";
 import IGraphic from "./IGraphic";
 import BoundingBox from "./BoundingBox";
-import Point from "./Point";
+import Point from "../Point";
 
 export default class Curve implements IGraphic {
     public id: string;
